Add !package keyword for new rule targeting (#980)
* Add !package keyword for new rule targeting

This new keyword helps target placement of new rules near the package()
declaration in a file. This is useful for adding new load() and
license() statements since it keeps them near the top since they are
globally significant rules.

* Restructured the scanning code.

Simplified the code to use the original loop and just check inside that
loop.