Add remove_if_equal command to buildozer (#1099)
* Add remove_if_equal command to buildozer

I've left open the structure to removing more complex expressions, but
trigger an error in all but the more basic cases.  I wasn't sure what
sort of behavior this type of command should have in more complex cases.

I generate a comparison function up-front, which does a lot of virtual
switching on types as there didn't seem to be a way to genericly compare
expressions.  I've used getAttrValueExpr, however, this does make for a
little weirdness in trying to handle the case of strings versus
identifiers.  The test case with `"True"` versus `True` tries to
demonstrate why identifiers are compared to both string and identifiers.

I assume this will seem incomplete in some respects, but I was putting
this up for initial feedback on the approach as well as what would be
expected in the test coverage.

Fixes #105